PSVAP/requirements.txt
PySide6==6.6.0
pyvista==0.44.0
pyvistaqt==0.11.0
numpy==1.26.4
scipy==1.12.0
MDAnalysis==2.7.0
biopython==1.83
# rdkit-pypi==2024.3.1
gemmi==0.6.6
freesasa==2.2.1
pyparsing==3.1.2
RestrictedPython==7.0
scikit-learn==1.4.0
matplotlib==3.8.0
colorcet==3.0.0
mordredcommunity==2.0.7
parmed==4.2.2
pdbfixer==1.12.0
griddataformats==1.0.1
pygments==2.17.0

